Implement fallback uncaught exception handler
Currently taskflow is not exception safe as exceptions thrown from user
tasks will break invariants required by the library and most likely
result in deadlocks. This situation can be improved by providing a
fallback exception handler which is called for any exceptions
originating from user tasks. This way exceptions are processed in a
well-defined way.

The API explicitly disclaims guarantee to provide the same exact way to
handle exceptions in future versions of the library, so this change does
not reduce freedom to improve exception handling in the future.